[ADD]: Auth 생성 LoginAPI Auth class 이동
