+++ IOS_Study/Screen/Login/Controller/LoginVC.swift
@@ -50,7 +50,7 @@ class LoginVC: UIViewController {
         UserDefaults.standard.set("jwtToken", forKey: "jwtToken")
         NotificationCenter.default.post(name: .authStateDidChange, object: nil)
         //showAlert(title: "로그인 실패", message: "회원정보가 없습니다")
-        //LoginAPI.requestLogin(params: login)
+        //Auth.requestLogin(params: login)
         print(login)
     }
     
@@ -175,36 +175,3 @@ extension LoginVC{
 }
 
 // MARK: function for login
-class LoginAPI{
-    static let url = "url"
-    
-    //로그인 요청
-    static func requestLogin(params login:Login){
-        let headers: HTTPHeaders = []
-        // 서버 resonpse 내용에 따라 코드 수정
-        AF.request(url, method: .post, parameters: login, encoder: JSONParameterEncoder.default, headers: headers)
-            .responseData { response in switch response.result {
-            case .success(let data):
-                let jwtToken = LoginAPI.parseData(data)
-                //자동 로그인을 위한 jwt 저장
-                UserDefaults.standard.set(jwtToken, forKey: "jwtToken")
-                NotificationCenter.default.post(name: .authStateDidChange, object: nil)
-            case.failure(let error):
-                print("error: \(error)")
-            }
-        }
-    }
-    
-    // decode data
-    static func parseData(_ data:Data)->String{
-        let decoder = JSONDecoder()
-        do{
-            let response = try decoder.decode(LoginResponse.self, from: data)
-            let jwtToken = response.jwtToken
-            return jwtToken
-        }catch let error{
-            print("error--->\(error)")
-            return ""
-        }
-    }
-}
